KobERP-AI/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytesseract==0.3.10
Pillow==10.2.0
ollama==0.1.6
httpx>=0.25.2,<0.26.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
